refactor(playgrounds): share command HUD UI
Extract the working command console HUD layout and scrolling behavior into the new `game-commands` utility crate so playgrounds use one Bevy plugin for command output UI. This fixes the SBS playground console cutoff by reusing the SDF playground's scrollable viewport structure instead of maintaining a divergent local implementation.

Both the SDF common playground and Chico SBS trees playground now keep only their playground-specific status text while delegating HUD setup, wheel scrolling, PageUp/PageDown scrolling, and shared console components to `GameCommandUiPlugin`.

Validation:

- `cargo test -p game-commands -p sdf-common-playground -p chico-sbs-trees-playground`

## Agent Dialogue

- **Developer:** Noted that the SBS trees playground UI cuts off command output, while the SDF common playground UI works, and asked to extract the working UI into `util/game-commands` as a plugin used by both crates.
- **Agent:** Created the `game-commands` crate, added `GameCommandUiPlugin` with the working scrollable HUD layout, wired both playgrounds to use it, trimmed their local UI modules to status/console text updates, and verified targeted tests.
