+++ maybraid/chico/sbs-trees-playground/Cargo.toml
@@ -13,6 +13,7 @@ rust-version = { workspace = true }
 bevy = { workspace = true }
 chunk = { workspace = true }
 clap = { workspace = true }
+game-commands = { workspace = true }
 log = { workspace = true }
 render-item = { workspace = true }
 sdf = { workspace = true }
