+++ maybraid/sdf/common-playground/Cargo.toml
@@ -12,6 +12,7 @@ rust-version = { workspace = true }
 [dependencies]
 bevy = { workspace = true }
 chunk = { workspace = true }
+game-commands = { workspace = true }
 log = { workspace = true }
 render-item = { workspace = true }
 sdf = { workspace = true }
