+++ maybraid/sdf/common-playground/src/lib.rs
@@ -8,22 +8,23 @@ pub mod commands;
 mod ground;
 mod input;
 mod preview;
-mod startup;
 pub mod primitive;
+mod startup;
 mod ui;
 
 pub use camera::CameraController;
 pub use commands::{PlaygroundCommand, PlaygroundCommandsPlugin, PLAYGROUND_CLI_NAME};
 pub use ground::PlaygroundSettings;
-pub use startup::PendingStartupCommand;
 pub use preview::{PreviewConfig, SdfPreviewRoot};
+pub use startup::PendingStartupCommand;
 
 use bevy::prelude::*;
-use ground::setup_ground;
 use commands::root::react_playground_command_root;
+use game_commands::ui::GameCommandUiPlugin;
+use ground::setup_ground;
 use preview::{keyboard_preview, sync_sdf_preview};
-use render_item::{mesh::fetch_meshes, mesh::handle::MeshHandle, render_items};
 use primitive::{PlaygroundPrimitive, PlaygroundRenderItem};
+use render_item::{mesh::fetch_meshes, mesh::handle::MeshHandle, render_items};
 
 /// Brown-ish default material for SDF previews (similar stick/trunk tone in objects playground).
 #[derive(Resource, Clone)]
@@ -37,22 +38,18 @@ impl Plugin for SdfCommonPlaygroundPlugin {
 			.init_resource::<PreviewConfig>()
 			.init_resource::<PendingStartupCommand>()
 			.add_plugins(PlaygroundCommandsPlugin)
+			.add_plugins(GameCommandUiPlugin { config: ui::ui_config() })
 			.add_plugins(startup::StartupPlugin)
-			.add_observer(ui::on_console_viewport_scroll)
-			.add_plugins(bevy::pbr::MaterialPlugin::<checkerboard_material::CheckerboardMaterial>::default())
+			.add_plugins(
+				bevy::pbr::MaterialPlugin::<checkerboard_material::CheckerboardMaterial>::default(),
+			)
 			.init_resource::<input::TypedCommandLine>()
 			.init_resource::<input::TextEntryFocus>()
 			.init_resource::<input::CommandConsoleOutput>()
 			.init_resource::<input::CommandHistory>()
 			.add_systems(
 				Startup,
-				(
-					camera::setup_camera,
-					setup_lighting,
-					setup_ground,
-					setup_preview_material,
-					ui::setup_debug_ui,
-				),
+				(camera::setup_camera, setup_lighting, setup_ground, setup_preview_material),
 			)
 			.add_systems(
 				Update,
@@ -61,8 +58,6 @@ impl Plugin for SdfCommonPlaygroundPlugin {
 					keyboard_preview,
 					input::toggle_text_entry_focus,
 					input::capture_command_line_input,
-					ui::send_console_ui_scroll_events,
-					ui::scroll_console_viewport_keyboard,
 					sync_sdf_preview.after(react_playground_command_root),
 					ui::update_debug_ui,
 					render_items::<PlaygroundRenderItem<StandardMaterial>>,
