+++ util/game-commands/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "game-commands"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+publish = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+bevy = { workspace = true }
+
+[lints]
+workspace = true
