+++ util/game-commands/README.md
@@ -1,0 +1,4 @@
+# `game-commands`
+
+A crate filled with useful helpers for building game command plugins in Bevy. 
+
