+++ util/game-commands/src/lib.rs
@@ -1,0 +1,3 @@
+//! Helpers for building in-game command UIs in Bevy.
+
+pub mod ui;
